src/requirements.txt
uvicorn==0.18.3
fastapi==0.85.0
sqlalchemy==1.4.41
psycopg2-binary==2.9.4
passlib[bcrypt]==1.7.4
pyjwt==2.5.0